reset tally reporter on each tick
currently tally reports all the metrics which existed during the
exporters runtime, even if they are not updated anymore. This change
reset tally each tick, causing only new metrics to be available for
exporter
