+++ main.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -21,7 +21,7 @@ import (
 
 func main() {
 	var wg sync.WaitGroup
-	mInfo := metrics.NewMetricReporter()
+	mInfo := metrics.NewMetricProvider()
 	defer mInfo.Close()
 
 	go mInfo.ServeHttp()
